Crear modulos independientes para el modelo
