+++ i18n/es.po
@@ -7,7 +7,7 @@ msgstr ""
 "Project-Id-Version: Prescrypto Server 10.0\n"
 "Report-Msgid-Bugs-To: \n"
 "POT-Creation-Date: 2017-09-25 21:10+0000\n"
-"PO-Revision-Date: 2017-09-25 14:11-0700\n"
+"PO-Revision-Date: 2017-09-25 17:54-0700\n"
 "Last-Translator: \n"
 "Language-Team: \n"
 "Language: es\n"
@@ -47,7 +47,7 @@ msgstr "Direcciónes"
 #: selection:partner.address,address_type:0
 #, python-format
 msgid "Administrative Address"
-msgstr " Administrativo"
+msgstr "Domicilio Administrativo"
 
 #. module: campos_clientes_vittal
 #: code:addons/campos_clientes_vittal/models/models.py:42
