+++ models/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from . import models
+from user_client import user_client
+from partner_member import partner_member
+from partner_address import partner_address
