+++ models/user_client.py
@@ -107,69 +107,3 @@ class user_client(models.Model):
         for record in self:
             if record.protected_home_auto or record.protected_home_user:
                 record.protected_home = True
-
-
-class partner_member(models.Model):
-    _name = 'partner.member'
-    _inherit = 'res.partner'
-
-    # número de socio
-    member_id = fields.Char(
-        string="Associate ID", help="Unique identifier for associates")
-
-    # padre del miembro
-    parent_id = fields.Many2one(
-        'res.partner', string="Parent", ondelete="set null")
-
-    # titular del grupo
-    is_owner = fields.Boolean(string="Owner", help="Is this the owner")
-
-    # condición de grupo
-    relationship = fields.Selection(
-        string="Relationship",
-        help="Type of member",
-        selection=[("1", "Spouse"), ("3", "Offspring"), ("4", "Other family"),
-                   ("9", "Not family")])
-
-    # fecha de nacimiento
-    birthday = fields.Date(string="Birthday", help="Date of birth")
-
-
-class partner_address(models.Model):
-    _name = 'partner.address'
-    _inherit = 'res.partner'
-
-    # padre de la dirección
-    parent_id = fields.Many2one(
-        'res.partner', string="Parent", ondelete="set null")
-
-    # tipos de dirección
-    address_type = fields.Selection(
-        string="Address Type",
-        help="Group addresses",
-        selection=[('admin', odoo._('Administrative Address')),
-                   ('fiscal', odoo._('Fiscal Address')),
-                   ('coverage',
-                    odoo._('Coverage Address')), ('attention',
-                                                  odoo._('Support Address'))])
-
-    # entre calles
-    cross_street = fields.Char(
-        string="Cross Streets", help="Nearest intersection")
-
-    # referencias
-    references = fields.Char(
-        string="References", help="Nearby landmarks or reference points")
-
-    # fachada
-    exterior = fields.Char(
-        string="Exterior", help="Description of exterior of address")
-
-    # características especiales
-    details = fields.Char(
-        string="Details", help="Special details or characteristics")
-
-    # establecer como domicilio de atención default
-    default_address = fields.Boolean(
-        string="Default Address", help="Default address for the group")
-
