+++ views/view_partner_form.xml
@@ -1,11 +1,5 @@
 <odoo>
     <data>
-        <template id="assets_backend" name="campos_clientes_vittal assets" inherit_id="web.assets_backend">
-            <xpath expr="." position="inside">
-                <link rel="stylesheet" href="/campos_clientes_vittal/static/src/css/custom.css"/>
-            </xpath>
-        </template>
-
         <record model="ir.ui.view" id="view_partner_form">
             <field name="name">res.partner.form.inherit</field>
             <field name="model">res.partner</field>
